compiler: detect and re-write switches
The Go compiler produces code that makes extremely heavy use of
C-style switch code, with some runtime functions (like `unicode.init`)
ending up with thousands of clauses.  While using continuations to
implement those blocks is natural, it ends up being extremely slow due
to the sheer number of continuations that have to be allocated and
GC'd.  This change makes it so that we detect that type of code using
heuristics and then re-write it into `letrec` code that uses plain
functions for flow control.